requirements.txt
SpeechRecognition
NOTE: for windows, if you're using python 3.7, you need to install precompiled binaries for compatibility sake, so install
pipwin first
pyaudio
pydub
ffmpeg

pocketsphinx
 google-cloud-speech (only for google cloud)

API:
pyowm
wikipedia
omdb
googletrans
gmail
pip3 install git+https://github.com/alberanid/imdbpy
googlesearch
pyserial

 TTS:
pyttsx3 (requires sudo apt-get update && sudo apt-get install espeak)y

gTTS